Adding in the logic to get the queries working

+++ stores/src/main/java/org/skye/stores/archive/localfs/LocalFSArchiveStore.java
@@ -98,7 +98,7 @@ public class LocalFSArchiveStore implements ArchiveStore, QueryableStore {
     public Optional<InputStream> getStream(ObjectMetadata metadata) {
         try {
             if (metadata.getArchiveContentBlock(this.getArchiveStoreDefinition().get().getId()).isPresent()) {
-                InputStream is = new FileInputStream(getSimpleObjectPath(metadata.getArchiveContentBlock(this.getArchiveStoreDefinition().get().getId()).get()));
+                InputStream is = new FileInputStream(getSimpleObjectPath(metadata.getArchiveContentBlock(this.getArchiveStoreDefinition().get().getId()).get(), metadata, false));
                 return Optional.of(is);
             } else return Optional.absent();
         } catch (FileNotFoundException e) {
@@ -112,7 +112,7 @@ public class LocalFSArchiveStore implements ArchiveStore, QueryableStore {
             if (metadata.getImplementation().equals(JDBCStructuredObject.class.getCanonicalName())) {
                 if (metadata.getArchiveContentBlock(this.getArchiveStoreDefinition().get().getId()).isPresent()) {
 
-                    UpdateableDataContext dataContext = createCsvDataContext(getSimpleObjectPath(metadata.getArchiveContentBlock(this.getArchiveStoreDefinition().get().getId()).get()));
+                    UpdateableDataContext dataContext = createCsvDataContext(getSimpleObjectPath(metadata.getArchiveContentBlock(this.getArchiveStoreDefinition().get().getId()).get(), metadata, false));
                     SimpleObject simpleObject = new JDBCStructuredObject(dataContext);
                     return Optional.of(simpleObject);
                 } else return Optional.absent();
@@ -141,8 +141,7 @@ public class LocalFSArchiveStore implements ArchiveStore, QueryableStore {
     @Override
     public void destroy(ObjectMetadata om) {
         if (om.getArchiveContentBlock(this.getArchiveStoreDefinition().get().getId()).isPresent()) {
-            getSimpleObjectPath(om.getArchiveContentBlock(this.getArchiveStoreDefinition().get().getId()).get()).delete();
-
+            getSimpleObjectPath(om.getArchiveContentBlock(this.getArchiveStoreDefinition().get().getId()).get(), om, false).delete();
         }
     }
 
@@ -154,25 +153,35 @@ public class LocalFSArchiveStore implements ArchiveStore, QueryableStore {
         return tmpPath;
     }
 
-    public File getSimpleObjectPath(ArchiveContentBlock acb) {
-        File simpleObjectDir = new File(getLocalPath() + "/" + acb.getId());
-        log.info("Storing object with ACB [" + getLocalPath() + "/" + acb.getId() + "]");
+    public File getSimpleObjectPath(ArchiveContentBlock acb, ObjectMetadata om, boolean isNew) {
+        String fileName = getLocalPath() + "/" + acb.getId() + "/" + om.getPath() + ".csv";
+        File simpleObjectDir = new File(fileName);
+        log.info("Storing object with ACB [" + getLocalPath() + "/" + acb.getId() + "/" + om.getPath() + "]");
 
-        if (simpleObjectDir.exists())
-            throw new SkyeException("Simple object already archived? " + acb);
-        simpleObjectDir.mkdirs();
-        simpleObjectDir.delete();
+        if (isNew) {
+            if (simpleObjectDir.exists())
+                throw new SkyeException("Simple object already archived? " + acb);
+            simpleObjectDir.mkdirs();
+            simpleObjectDir.delete();
+        } else if (!simpleObjectDir.exists()) {
+            throw new SkyeException("Unable to find simple object for acb " + acb);
+        }
         return simpleObjectDir;
     }
 
-    public File getTempSimpleObjectPath(ArchiveContentBlock acb) {
-        File simpleObjectDir = new File(getTempPath() + "/" + acb.getId());
-        log.info("Storing temp object with ACB [" + getTempPath() + "/" + acb.getId() + "]");
+    public File getTempSimpleObjectPath(ArchiveContentBlock acb, ObjectMetadata om, boolean isNew) {
+        String fileName = getTempPath() + "/" + acb.getId() + "/" + om.getPath() + ".csv";
+        File simpleObjectDir = new File(fileName);
+        log.info("Storing temp object with ACB [" + getTempPath() + "/" + acb.getId() + "/" + om.getPath() + "]");
 
-        if (simpleObjectDir.exists())
-            throw new SkyeException("Simple object already in temp? " + acb);
-        simpleObjectDir.mkdirs();
-        simpleObjectDir.delete();
+        if (isNew) {
+            if (simpleObjectDir.exists())
+                throw new SkyeException("Simple object already in temp? " + acb);
+            simpleObjectDir.mkdirs();
+            simpleObjectDir.delete();
+        } else if (!simpleObjectDir.exists()) {
+            throw new SkyeException("Unable to find simple object for acb " + acb);
+        }
         return simpleObjectDir;
     }
 
@@ -182,13 +191,12 @@ public class LocalFSArchiveStore implements ArchiveStore, QueryableStore {
         List<StructuredObject> contextObjects = context.resolveObjects(this);
         List<DataContext> dataContexts = new ArrayList<>();
         for (StructuredObject obj : contextObjects) {
-            dataContexts.add(((JDBCStructuredObject) obj).getDataContext());
+            DataContext dc = ((JDBCStructuredObject) obj).getDataContext();
+            dataContexts.add(dc);
         }
 
         // Build a composite context
         DataContext compositeDataContext = DataContextFactory.createCompositeDataContext(dataContexts);
-
-        QueryResultStructuredObject result = new QueryResultStructuredObject(compositeDataContext.executeQuery(query));
-        return null;  //To change body of implemented methods use File | Settings | File Templates.
+        return new QueryResultStructuredObject(compositeDataContext.executeQuery(query));
     }
 }
